requirements.txt
yahoo_fin
panel
seaborn
plotly
finnhub-python
pandas
folium
numba
holoviews
scipy
hvplot
openpyxl